requirements.txt
accelerate==0.33.0
aiofiles==23.2.1
aiohttp==3.9.5
aioprometheus==23.12.0
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
argcomplete==3.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.30.1
azure-identity==1.16.0
backcall==0.2.0
bitsandbytes==0.43.3
Brotli==1.0.9
cachetools==5.3.3
Cerberus==1.3.5
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpickle==2.2.1
cmake==3.30.2
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.1.1
coverage==7.2.7
cryptography==42.0.5
cssutils==2.11.1
cupy-cuda12x==12.1.0
cycler==0.12.1
cytoolz==0.12.2
dask==2023.4.1
datasets==2.20.0
debugpy==1.8.2
decorator==5.1.1
deepspeed==0.15.0
diffusers==0.29.2
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
dpcpp-cpp-rt==2024.2.1
einops==0.8.0
eval_type_backport==0.2.0
evaluate==0.4.2
exceptiongroup==1.2.1
executing==2.0.1
fairscale==0.4.13
fastapi==0.112.2
fastjsonschema==2.20.0
fastrlock==0.8.2
ffmpeg==1.4
filelock==3.13.1
flash-attn==2.5.9.post1
flatbuffers==24.3.25
fonttools==4.53.1
Fraction==2.2.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.2
greenlet==3.0.3
h11==0.14.0
h5py==3.11.0
hjson==3.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.7
igraph==0.11.4
imagecodecs==2021.8.26
imageio==2.33.1
imageio-ffmpeg==0.5.1
importlib-metadata==7.0.1
importlib_resources==6.4.0
iniconfig==2.0.0
intel-cmplr-lib-rt==2024.2.1
intel-cmplr-lib-ur==2024.2.1
intel-cmplr-lic-rt==2024.2.1
intel-opencl-rt==2024.2.1
intel-openmp==2024.2.1
intel-sycl-rt==2024.2.1
ipykernel==6.29.5
ipython==8.12.2
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.2
llvmlite==0.41.1
locket==1.0.0
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2024.2.1
mkl-fft==1.3.6
mkl-include==2022.2.1
mkl-random==1.2.2
mkl-service==2.4.0
more-itertools==10.3.0
mpmath==1.3.0
msal==1.28.0
msal-extensions==1.1.0
msccl @ git+https://github.com/microsoft/msccl-tools.git@030a750fc56e03fb863ffe48268672d84d4cdc7b
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
numba==0.58.1
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.6.1
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
opencensus-ext-logging==0.1.1
opencv-python==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.0.3
parso==0.8.4
partd==1.4.1
pathtools==0.1.2
peft==0.12.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
portalocker==2.8.2
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyaml==24.7.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
pynvml==11.5.0
pyparsing==3.1.2
PySocks==1.7.1
pytest==7.4.3
pytest-mpi==0.6
python-dateutil==2.9.0
python-dotenv==1.0.1
pytorch-lightning==1.9.5
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==26.0.3
quantile-python==1.1
ray==2.35.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.3
scikit-image==0.20.0
scikit-learn==1.3.2
scipy==1.9.1
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.10.0
seqeval==1.2.2
setproctitle==1.3.3
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.31
stack-data==0.6.2
starlette==0.38.4
svgpathtools==1.6.1
svgwrite==1.4.3
sympy==1.12
tabulate==0.9.0
tbb==2021.12.0
tensor-parallel==1.2.4
texttable==1.7.0
threadpoolctl==3.5.0
tifffile==2021.11.2
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.4.0
torch-ort==1.17.0
torch-tools==0.1.5
torchaudio==2.4.0
torchdata==0.7.1
torchmetrics==1.2.0
torchsnapshot==0.1.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers @ git+https://github.com/leloykun/transformers.git@dae439cb93e7c2ef61943554b7a2f4590c2118cd
triton==3.0.0
trl==0.9.6
tutel @ https://github.com/microsoft/tutel/archive/refs/tags/v0.3.1.tar.gz
typing_extensions==4.11.0
tyro==0.8.5
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
visdom==0.2.4
wandb==0.17.4
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.0.1
wget==3.2
widgetsnbextension==4.0.13
xxhash==3.4.1
yarl==1.9.4
z3-solver==4.13.0.0
zipp==3.17.0
torch-fidelity==0.3.0
lpips==0.1.4
distributed==2023.4.1
dreamsim==0.2.1
evo==1.31.1
pytorch-msssim==1.0.0
